Future-proof build script + tests (#71)
This patch leverages the new RGI_Emoji property to simplify and future-proof the build script. It also changes both the build script and the tests to avoid hardcoding any particular Unicode version in their source, so that updating emoji-regex for future Unicode Standard releases is simpler than ever: it can be done just by adding the new `@unicode/unicode-*` devDependency to `package.json`.